movie-mas/requirements.txt
langchain-groq>=0.1.9
langgraph>=0.2.28
streamlit>=1.38.0
python-dotenv>=1.0.1
